internal/appsec: add stubs for waf_disabled.go
