+++ internal/appsec/waf/types.go
@@ -5,10 +5,29 @@
 
 package waf
 
-import "fmt"
+import (
+	"fmt"
+	"sync/atomic"
+)
 
 // RunError the WAF can return when running it.
 type RunError int
+
+// AtomicU64 can be used to perform atomic operations on an uint64 type
+type AtomicU64 uint64
+
+// RulesetInfo stores the information - provided by the WAF - about WAF rules initialization.
+type RulesetInfo struct {
+	// Number of rules successfully loaded
+	Loaded uint16
+	// Number of rules which failed to parse
+	Failed uint16
+	// Map from an error string to an array of all the rule ids for which
+	// that error was raised. {error: [rule_ids]}
+	Errors map[string]interface{}
+	// Ruleset version
+	Version string
+}
 
 // Errors the WAF can return when running it.
 const (
@@ -38,4 +57,14 @@ func (e RunError) Error() string {
 	default:
 		return fmt.Sprintf("unknown waf error %d", e)
 	}
+}
+
+// Add atomically sums the current atomic value with the provided value `v`.
+func (a *AtomicU64) Add(v uint64) {
+	atomic.AddUint64((*uint64)(a), v)
+}
+
+// Inc atomically increments the atomic value by 1
+func (a *AtomicU64) Inc() {
+	atomic.AddUint64((*uint64)(a), 1)
 }
